requirements.txt
# Core AI/ML Dependencies
langchain>=0.1.0
langchain-google-genai>=0.0.6
langgraph>=0.0.20
google-generativeai>=0.3.0

# Web Frameworks
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
streamlit>=1.28.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# API & Networking
requests>=2.31.0
python-multipart>=0.0.6

# Environment Management
python-dotenv>=1.0.0
